requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Auth + persistence
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
asyncpg>=0.29.0
email-validator>=2.0.0

# Monetization
stripe>=8.0.0

# Migrations
alembic>=1.13.0
SQLAlchemy>=2.0.0

# Security / Linting (dev only — not required at runtime)
# bandit>=1.7.0
# flake8>=7.0.0
